+++ common/src/main/java/revamped_phantoms/Utils.java
@@ -5,11 +5,12 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobSpawnType;
 import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.level.ServerLevelAccessor;
+import net.minecraft.world.phys.AABB;
 
 import java.util.Random;
 
 public class Utils {
     public static boolean shouldSpawn(EntityType<? extends Monster> type, ServerLevelAccessor level, MobSpawnType reason, BlockPos pos, Random random) {
-        return Monster.checkMonsterSpawnRules(type,level,reason,pos,random) && level.getSkyDarken() >= 4 && level.canSeeSky(pos);
+        return Monster.checkMonsterSpawnRules(type,level,reason,pos,random) && !level.collidesWithSuffocatingBlock(null, new AABB(pos, pos.atY(level.getMaxBuildHeight()).offset(1,0,1)));
     }
 }