+++ common/src/main/java/revamped_phantoms/mixin/SpawnPlacementsMixin.java
@@ -2,8 +2,8 @@ package revamped_phantoms.mixin;
 
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.SpawnPlacements;
+import net.minecraft.world.level.levelgen.Heightmap;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import revamped_phantoms.Utils;
@@ -11,18 +11,9 @@ import revamped_phantoms.Utils;
 @Mixin(SpawnPlacements.class)
 public class SpawnPlacementsMixin {
 
-    @Unique
-    private static EntityType revamped_phantoms_lastEntity;
-
-    @ModifyVariable(method="register", at=@At("HEAD"), argsOnly=true)
-    private static EntityType revamped_phantoms_modifyPhantomSpawning(EntityType entityType) {
-        revamped_phantoms_lastEntity = entityType;
-        return entityType;
-    }
-
-    @ModifyVariable(method="register", at=@At("HEAD"), argsOnly=true)
-    private static SpawnPlacements.SpawnPredicate revamped_phantoms_modifyPhantomSpawning(SpawnPlacements.SpawnPredicate spawnPredicate) {
-        if (revamped_phantoms_lastEntity == EntityType.PHANTOM) {
+    @ModifyVariable(method="register",at=@At("HEAD"), argsOnly = true)
+    private static SpawnPlacements.SpawnPredicate revamped_phantoms_modifyPhantomSpawning(SpawnPlacements.SpawnPredicate spawnPredicate, EntityType entityType, SpawnPlacements.Type placement, Heightmap.Types heightMap, SpawnPlacements.SpawnPredicate spawnPredicateUnused) {
+        if (entityType == EntityType.PHANTOM) {
             return Utils::shouldSpawn;
         }
         return spawnPredicate;